requirements.txt
kafka-python
flask
flask-cors
matplotlib